+++ src/core/frontend_app/index.ts
@@ -14,15 +14,25 @@ export {
 export {
   DEFAULT_FRONTEND_CLIENT_ENTRY_KEY,
   DEFAULT_FRONTEND_DEFERRED_CLIENT_ENTRY_KEY,
+  DEFAULT_FRONTEND_GLOBAL_CLIENT_ENTRY_PATTERNS,
   DEFAULT_FRONTEND_GLOBAL_STYLE_PATTERNS,
   DEFAULT_FRONTEND_GLOBAL_STYLE_RULE_KEY,
   DEFAULT_FRONTEND_IGNORED_SOURCE_PATTERNS,
   DEFAULT_FRONTEND_PUBLIC_DIR,
   DEFAULT_FRONTEND_RUNTIME_SOURCE_PATTERNS,
   DEFAULT_FRONTEND_SOURCE_DIR,
+  DEFAULT_FRONTEND_SSR_MAP_EXPORT,
+  DEFAULT_FRONTEND_SSR_MATCHED_EXPORT,
   DEFAULT_FRONTEND_SSR_PAGE_PATTERNS,
+  DEFAULT_FRONTEND_SSR_RESOLVER_EXPORT,
+  DEFAULT_FRONTEND_SSR_ROOT_EXPORT,
+  DEFAULT_FRONTEND_SSR_ROOT_MODULE_EXPORT,
   DEFAULT_FRONTEND_SSR_RULE_KEY,
 } from "./defaults.js";
+export {
+  resolveConfiguredFrontendGlobalClientEntries,
+  resolveFrontendGlobalClientEntries,
+} from "./global.js";
 export {
   quarantineUnwritableOutputDir,
 } from "./hygiene.js";
@@ -40,11 +50,15 @@ export {
   buildRelatedClientEntryMap,
 } from "./related.js";
 export {
+  prepareSsrNodeModules,
+} from "./node_modules.js";
+export {
   createFrontendBundlerRuntime,
 } from "./runtime.js";
 export {
   createReactSsrModuleMapRule,
   createSsrModuleMapRule,
+  resolveSsrModuleMapRuleOptions,
 } from "./ssr.js";
 export {
   createStaticAssetMiddleware,